--- FakeNFT/Models/Collection.swift
@@ -1,7 +1,0 @@
-import Foundation
-
-struct CollectionItem {
-     let imageStringURL: String
-     let title: String
-     let count: Int
- }
