+++ FakeNFT/Screens/Catalog/Services/CatalogProtocol.swift
