+++ FakeNFT/Screens/Catalog/Services/CollectionDataProvider.swift
@@ -1,4 +1,9 @@
-import Foundation 
+import Foundation
+
+enum ResultModel<T> {
+    case success(T)
+    case failure(Error)
+}
 
 final class CollectionDataProvider: CollectionDataProviderProtocol {
     private let networkClient: NetworkClient
@@ -7,15 +12,10 @@ final class CollectionDataProvider: CollectionDataProviderProtocol {
         self.networkClient = networkClient
     }
     
-    func getOrder(completion: @escaping (Result<OrderModel, Error>) -> Void) {
+    func getOrder(completion: @escaping (ResultModel<OrderModel>) -> Void) {
         let getOrderRequest = GetOrderRequest()
         networkClient.send(request: getOrderRequest, type: OrderModel.self) { result in
-            switch result {
-            case .success(let order):
-                completion(.success(order))
-            case .failure(let error):
-                completion(.failure(error))
-            }
+            self.handleResult(result, completion: completion)
         }
     }
     
@@ -31,39 +31,24 @@ final class CollectionDataProvider: CollectionDataProviderProtocol {
         }
     }
     
-    func getNFT(by id: String, completion: @escaping (Result<NFTModel, Error>) -> Void) {
+    func getNFT(by id: String, completion: @escaping (ResultModel<NFTModel>) -> Void) {
         let getNFTByIdRequest = GetNFTByIdRequest(id: id)
         networkClient.send(request: getNFTByIdRequest, type: NFTModel.self) { result in
-            switch result {
-            case .success(let nft):
-                completion(.success(nft))
-            case .failure(let error):
-                completion(.failure(error))
-            }
+            self.handleResult(result, completion: completion)
         }
     }
     
-    func getAuthor(by id: String, completion: @escaping (Result<AuthorModel, Error>) -> Void) {
+    func getAuthor(by id: String, completion: @escaping (ResultModel<AuthorModel>) -> Void) {
         let getUserByIdRequest = GetUserByIdRequest(id: id)
         networkClient.send(request: getUserByIdRequest, type: AuthorModel.self) { result in
-            switch result {
-            case .success(let user):
-                completion(.success(user))
-            case .failure(let error):
-                completion(.failure(error))
-            }
+            self.handleResult(result, completion: completion)
         }
     }
     
-    func getFavorites(completion: @escaping (Result<FavoritesModel, Error>) -> Void) {
+    func getFavorites(completion: @escaping (ResultModel<FavoritesModel>) -> Void) {
         let getFavoritesRequest = GetFavoritesRequest()
         networkClient.send(request: getFavoritesRequest, type: FavoritesModel.self) { result in
-            switch result {
-            case .success(let favoritesModel):
-                completion(.success(favoritesModel))
-            case .failure(let error):
-                completion(.failure(error))
-            }
+            self.handleResult(result, completion: completion)
         }
     }
     
@@ -76,6 +61,15 @@ final class CollectionDataProvider: CollectionDataProviderProtocol {
             case .failure(let error):
                 completion(.failure(error))
             }
+        }
+    }
+    
+    private func handleResult<T>(_ result: Result<T, Error>, completion: @escaping (ResultModel<T>) -> Void) {
+        switch result {
+        case .success(let value):
+            completion(.success(value))
+        case .failure(let error):
+            completion(.failure(error))
         }
     }
 }
